requirements.txt
rich==13.5.2
typer==0.9.0
python-dotenv==1.0.0
google-generativeai==0.3.1
openai==1.3.0
anthropic==0.5.0
ollama==0.1.6
requests==2.31.0
pathlib==1.0.1
colorama==0.4.6
pyyaml==6.0.1
tqdm==4.66.1 